Rename field-kit/core to field-kit/src.
