+++ packages/field-kit/jsconfig.json
@@ -1,5 +1,5 @@
 {
   "include": [
-    "./core/**/*"
+    "./src/**/*"
   ]
 }
