+++ packages/field-kit/src/utils/flattenEntity.js
